updated median function int typecasting removed, error.php added

+++ PHP/DbConnect.php
@@ -363,7 +363,7 @@ class Database {
 		}
 		else{
 			
-			$answer = $data[int(count($data)/2)];
+			$answer = $data[count($data)/2];
 			
 		}
